Changing code to use only one Utility class
